Use int arithmetic for IdLinkStore page-slice growth length
No functional change. The previous expression was already valid Go (make
accepts any integer-typed length), but computing the delta in int avoids
the inner uint64(len(...)) cast and reads more clearly.

+++ internal/core/linkstore.go
@@ -59,7 +59,7 @@ func (s *IdLinkStore[V]) Get(id uint64) *V {
 func (s *IdLinkStore[V]) getSlow(id uint64) *V {
 	page := id >> idLinkPageShift
 	if page >= uint64(len(s.pages)) {
-		s.pages = append(s.pages, make([][]*V, page+1-uint64(len(s.pages)))...)
+		s.pages = append(s.pages, make([][]*V, int(page)+1-len(s.pages))...)
 	}
 	if s.pages[page] == nil {
 		s.pages[page] = make([]*V, idLinkPageSize)
